requirements.txt
APScheduler==3.11.2
fpdf2==2.8.4
matplotlib==3.9.4
mplfinance==0.12.10b0
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.3
pytest==8.4.2
python-dotenv==1.2.1
seaborn==0.13.2
tabulate==0.9.0
yfinance==1.2.0
